rework on authentication logic in home pages
