+++ frontend/src/components/Home/HomeRoomCalendar.vue
@@ -1,0 +1,14 @@
+<template>
+  <v-sheet height="85vh">
+    <RoomCalendar />
+  </v-sheet>
+</template>
+
+<script>
+import RoomCalendar from '@/components/RoomCalendar.vue';
+
+export default {
+  name: 'HomeRoomCalendar',
+  components: { RoomCalendar },
+};
+</script>
