+++ frontend/src/pages/Home.vue
@@ -1,11 +1,6 @@
 <template>
   <div class="home">
-    <HomeHeader
-      :loading="loading"
-      :loading-out="false"
-      @login="login"
-      @logout="logout"
-    />
+    <HomeHeader />
     <v-container>
       <RoomMap />
     </v-container>
@@ -15,56 +10,11 @@
 <script>
 import HomeHeader from '@/components/Home/HomeHeader.vue';
 import RoomMap from '@/components/Home/RoomMap.vue';
-import { apiLogin } from '@/api/auth.api';
-import { mapActions } from 'vuex';
 export default {
   name: 'Home',
   components: {
     HomeHeader,
     RoomMap,
-  },
-  data: () => ({
-    loading: false,
-    loadingOut: false,
-  }),
-  mounted() {
-    const fenixCode = this.$route.query.code;
-    if (fenixCode) this.authBackend(fenixCode);
-  },
-  methods: {
-    login() {
-      window.location = `${import.meta.env.VITE_FENIX_BASE_URL}oauth/userdialog?client_id=${import.meta.env.VITE_FENIX_CLIENT_ID}&redirect_uri=${import.meta.env.VITE_FENIX_REDIRECT_URL}`;
-    },
-    //only runs after fenix login
-    async authBackend(code) {
-      this.loading = true;
-      try {
-        const { data } = await apiLogin(code);
-        if (data.jwt) {
-          localStorage.setItem('token', data.jwt);
-          this.loginUser(data.user);
-        }
-      } catch (e) {
-        //print exception
-        console.log(e);
-        this.$notify({
-          type: 'error',
-          title: 'Unauthorized user',
-          text: "You don't have permission to access this ",
-          duration: -1,
-        });
-      }
-      this.loading = false;
-      this.$router.push('login');
-    },
-    logout() {
-      this.loadingOut = true;
-      localStorage.removeItem('token');
-      this.logoutUser();
-      this.loadingOut = false;
-      window.open('https://fenix.tecnico.ulisboa.pt/logout', '_blank').focus();
-    },
-    ...mapActions('user', ['loginUser', 'logoutUser']),
-  },
+  }
 };
 </script>
