+++ frontend/src/pages/Printing.vue
@@ -1,11 +1,6 @@
 <template>
   <div class="home">
-    <HomeHeader
-      :loading="loading"
-      :loading-out="false"
-      @login="login"
-      @logout="logout"
-    />
+    <HomeHeader />
     <v-container>
       <HomePrinting />
     </v-container>
@@ -23,44 +18,6 @@ export default {
   components: {
     HomeHeader,
     HomePrinting,
-  },
-  data: () => ({
-    loading: false,
-    loadingOut: false,
-  }),
-  mounted() {
-    const fenixCode = this.$route.query.code;
-    if (fenixCode) this.authBackend(fenixCode);
-  },
-  methods: {
-    //only runs after fenix login
-    async authBackend(code) {
-      this.loading = true;
-      try {
-        const { data } = await apiLogin(code);
-        if (data.jwt) {
-          localStorage.setItem('token', data.jwt);
-          this.loginUser(data.user);
-        }
-      } catch (e) {
-        console.log(e);
-        this.$notify({
-          type: 'error',
-          title: 'Unauthorized user',
-          text: "You don't have permission to access this ",
-          duration: -1,
-        });
-      }
-      this.loading = false;
-      this.$router.push('login');
-    },
-    logout() {
-      this.loadingOut = true;
-      localStorage.removeItem('token');
-      this.logoutUser();
-      this.loadingOut = false;
-    },
-    ...mapActions('user', ['loginUser', 'logoutUser']),
-  },
+  }
 };
 </script>
